Fixed reveal closing randomly all .flex-videos from document
In the context of the function, this refers to document, so when you do $(document).find('.flex-video') and then set the iframe src to about:blank, all videos on the website are dying.

The huge problem is that this method is also called by alert-boxes, dropdowns and so on.
(cherry picked from commit b4bd0f47438205c25475849f4e58fb48d990ca3a)